pyproject.toml
[tool.poetry]
name = "let'sshop"
version = "0.1.0"
description = ""
authors = ["Mrudhulraj <mrudhulraj.n@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
Pillow = "^9.1.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
